refactor: Split DriveFileManagerConstants and RealmSchemaVersion
