+++ kDriveCore/Data/Cache/DriveFileManager/DriveFileManager.swift
@@ -28,178 +28,7 @@ import SwiftRegex
 public final class DriveFileManager: RealmAccessible {
     @LazyInjectService private var driveInfosManager: DriveInfosManager
 
-    /// Something to centralize schema versioning
-    enum RealmSchemaVersion {
-        /// Current version of the Upload Realm
-        static let upload: UInt64 = 18
-
-        /// Current version of the Drive Realm
-        static let drive: UInt64 = 9
-    }
-
-    // TODO: Make a type for the upload realm
-    public class DriveFileManagerConstants {
-        public let driveObjectTypes = [
-            File.self,
-            Rights.self,
-            FileActivity.self,
-            FileCategory.self,
-            FileConversion.self,
-            FileVersion.self,
-            FileExternalImport.self,
-            ShareLink.self,
-            ShareLinkCapabilities.self,
-            DropBox.self,
-            DropBoxCapabilities.self,
-            DropBoxSize.self,
-            DropBoxValidity.self
-        ]
-        private let fileManager = FileManager.default
-        public let rootDocumentsURL: URL
-        public let importDirectoryURL: URL
-        public let groupDirectoryURL: URL
-        public var cacheDirectoryURL: URL
-        public var tmpDirectoryURL: URL
-        public let openInPlaceDirectoryURL: URL?
-        public let rootID = 1
-        public let currentVersionCode = 1
-        public lazy var migrationBlock = { [weak self] (migration: Migration, oldSchemaVersion: UInt64) in
-            let currentUploadSchemaVersion = RealmSchemaVersion.upload
-
-            // Log migration on Sentry
-            SentryDebug.realmMigrationStartedBreadcrumb(
-                form: oldSchemaVersion,
-                to: currentUploadSchemaVersion,
-                realmName: "Upload"
-            )
-            defer {
-                SentryDebug.realmMigrationEndedBreadcrumb(
-                    form: oldSchemaVersion,
-                    to: currentUploadSchemaVersion,
-                    realmName: "Upload"
-                )
-            }
-
-            // Sanity check
-            guard oldSchemaVersion < currentUploadSchemaVersion else {
-                return
-            }
-
-            // Migration from version 2 to version 3
-            if oldSchemaVersion < 3 {
-                migration.enumerateObjects(ofType: UploadFile.className()) { _, newObject in
-                    newObject!["maxRetryCount"] = 3
-                }
-            }
-            // Migration to version 4 -> 7 is not needed
-            // Migration from version 7 to version 9
-            if oldSchemaVersion < 9 {
-                migration.deleteData(forType: DownloadTask.className())
-            }
-
-            // Migration from version 9 to version 10
-            if oldSchemaVersion < 10 {
-                migration.enumerateObjects(ofType: UploadFile.className()) { _, newObject in
-                    newObject!["conflictOption"] = ConflictOption.version.rawValue
-                }
-            }
-
-            if oldSchemaVersion < 12 {
-                migration.enumerateObjects(ofType: PhotoSyncSettings.className()) { _, newObject in
-                    newObject!["photoFormat"] = PhotoFileFormat.heic.rawValue
-                }
-            }
-
-            // Migration for Upload With Chunks
-            if oldSchemaVersion < 14 {
-                migration.deleteData(forType: UploadFile.className())
-            }
-
-            // Migration for UploadFile With dedicated fileProviderItemIdentifier and assetLocalIdentifier fields
-            if oldSchemaVersion < 15 {
-                migration.enumerateObjects(ofType: UploadFile.className()) { oldObject, newObject in
-                    guard let newObject else {
-                        return
-                    }
-
-                    // Try to migrate the assetLocalIdentifier if possible
-                    let type: String? = oldObject?["rawType"] as? String ?? nil
-
-                    switch type {
-                    case UploadFileType.phAsset.rawValue:
-                        // The object was from a phAsset source, the id has to be the `LocalIdentifier`
-                        let oldAssetIdentifier: String? = oldObject?["id"] as? String ?? nil
-
-                        newObject["assetLocalIdentifier"] = oldAssetIdentifier
-                        newObject["fileProviderItemIdentifier"] = nil
-
-                        // Making sure the ID is unique, and not a PHAsset identifier
-                        newObject["id"] = UUID().uuidString
-
-                    default:
-                        // We cannot infer anything, all to nil
-                        newObject["assetLocalIdentifier"] = nil
-                        newObject["fileProviderItemIdentifier"] = nil
-                    }
-                }
-            }
-        }
-
-        /// Path of the upload DB
-        public lazy var uploadsRealmURL = rootDocumentsURL.appendingPathComponent("uploads.realm")
-
-        public lazy var uploadsRealmConfiguration = Realm.Configuration(
-            fileURL: uploadsRealmURL,
-            schemaVersion: RealmSchemaVersion.upload,
-            migrationBlock: migrationBlock,
-            objectTypes: [DownloadTask.self,
-                          PhotoSyncSettings.self,
-                          UploadSession.self,
-                          UploadFile.self,
-                          UploadingChunkTask.self,
-                          UploadedChunk.self,
-                          UploadingSessionTask.self]
-        )
-
-        /// realm db used for file upload
-        public var uploadsRealm: Realm {
-            // Change file metadata after creation of the realm file.
-            defer {
-                // Exclude "upload file realm" and custom cache from system backup.
-                var metadata = URLResourceValues()
-                metadata.isExcludedFromBackup = true
-                do {
-                    try uploadsRealmURL.setResourceValues(metadata)
-                    try cacheDirectoryURL.setResourceValues(metadata)
-                } catch {
-                    DDLogError(error)
-                }
-            }
-
-            do {
-                return try Realm(configuration: uploadsRealmConfiguration)
-            } catch {
-                // We can't recover from this error but at least we report it correctly on Sentry
-                Logging.reportRealmOpeningError(error, realmConfiguration: uploadsRealmConfiguration)
-            }
-        }
-
-        init() {
-            @InjectService var pathProvider: AppGroupPathProvidable
-            groupDirectoryURL = pathProvider.groupDirectoryURL
-            rootDocumentsURL = pathProvider.realmRootURL
-            importDirectoryURL = pathProvider.importDirectoryURL
-            tmpDirectoryURL = pathProvider.tmpDirectoryURL
-            cacheDirectoryURL = pathProvider.cacheDirectoryURL
-            openInPlaceDirectoryURL = pathProvider.openInPlaceDirectoryURL
-
-            DDLogInfo(
-                "App working path is: \(fileManager.urls(for: .documentDirectory, in: .userDomainMask).first?.absoluteString ?? "")"
-            )
-            DDLogInfo("Group container path is: \(groupDirectoryURL.absoluteString)")
-        }
-    }
-
+    // TODO: Use DI
     public static let constants = DriveFileManagerConstants()
 
     private let fileManager = FileManager.default
